+++ example/android/app/src/main/java/com/mengtnt/export_video_frame_example/MainActivity.java
@@ -1,0 +1,7 @@
+package com.mengtnt.export_video_frame_example;
+
+import io.flutter.embedding.android.FlutterActivity;
+
+public class MainActivity extends FlutterActivity {
+  
+}