+++ example/lib/main.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 import 'dart:math';
 
+<<<<<<< HEAD
 import 'package:flutter/material.dart';
 import 'dart:async';
 
@@ -12,6 +13,51 @@ void main() {
   runApp( MyApp(images: <Image>[]));
 }
 
+=======
+/** 
+MIT License
+
+Copyright (c) 2019 mengtnt
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+*/
+
+import 'package:flutter/material.dart';
+import 'package:image_picker/image_picker.dart';
+import 'package:export_video_frame/export_video_frame.dart';
+
+void main() => runApp(MyApp());
+
+class MyApp extends StatelessWidget {
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: "Plugin Example App",
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: MyHomePage(images: <Image>[]),
+    );
+  }
+}
+>>>>>>> upstream/master
 
 class ImageItem extends StatelessWidget {
   ImageItem({required this.image}) : super(key: ObjectKey(image));
@@ -23,6 +69,7 @@ class ImageItem extends StatelessWidget {
   }
 }
 
+<<<<<<< HEAD
 class MyApp extends StatefulWidget {
   List<Image> images;
 
@@ -34,10 +81,23 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   String _platformVersion = 'Unknown';
+=======
+class MyHomePage extends StatefulWidget {
+  MyHomePage({Key? key, required this.images}) : super(key: key);
+
+  final List<Image> images;
+
+  @override
+  _MyHomePageState createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+>>>>>>> upstream/master
   var _isClean = false;
   final ImagePicker _picker = ImagePicker();
   Stream<File>? _imagesStream;
 
+<<<<<<< HEAD
   @override
   void initState() {
     super.initState();
@@ -72,11 +132,20 @@ class _MyAppState extends State<MyApp> {
   Future _getImagesByTime() async {
     final XFile? file = await _picker.pickVideo(
         source: ImageSource.gallery);
+=======
+  Future _getImagesByTime() async {
+    final PickedFile? file = await _picker.getVideo(
+        source: ImageSource.gallery, maxDuration: const Duration(seconds: 10));
+>>>>>>> upstream/master
     _imagesStream = ExportVideoFrame.exportImagesFromFile(
       File(file!.path),
       const Duration(milliseconds: 500),
       pi / 2,
+<<<<<<< HEAD
     ) ;
+=======
+    );
+>>>>>>> upstream/master
 
     setState(() {
       _isClean = true;
@@ -84,15 +153,25 @@ class _MyAppState extends State<MyApp> {
 
     _imagesStream!.listen((image) {
       setState(() {
+<<<<<<< HEAD
         widget.images.add(Image.file(File(image.path)));
+=======
+        widget.images.add(Image.file(image));
+>>>>>>> upstream/master
       });
     });
   }
 
   Future _getImages() async {
+<<<<<<< HEAD
     final XFile? file = await _picker.pickVideo(
         source: ImageSource.gallery,);
     var images = await ExportVideoFrame.exportImage(file!.path, 10, 100);
+=======
+    final PickedFile? file = await _picker.getVideo(
+        source: ImageSource.gallery, maxDuration: const Duration(seconds: 10));
+    var images = await ExportVideoFrame.exportImage(file!.path, 10, 0);
+>>>>>>> upstream/master
     var result = images.map((file) => Image.file(file)).toList();
     setState(() {
       widget.images.addAll(result);
@@ -101,8 +180,13 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future _getGifImages() async {
+<<<<<<< HEAD
     final XFile? file =
     await _picker.pickImage(source: ImageSource.gallery);
+=======
+    final PickedFile? file =
+        await _picker.getImage(source: ImageSource.gallery);
+>>>>>>> upstream/master
 
     var images = await ExportVideoFrame.exportGifImage(file!.path, 0);
     var result = images.map((file) => Image.file(file)).toList();
@@ -113,17 +197,29 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future _getImagesByDuration() async {
+<<<<<<< HEAD
     final XFile? pickedFile =
     await _picker.pickVideo(source: ImageSource.gallery);
     final File file = File(pickedFile!.path);
     var duration = Duration(seconds: 1);
     var image =
     await ExportVideoFrame.exportImageBySeconds(file, duration, 0);
+=======
+    final PickedFile? pickedFile =
+        await _picker.getImage(source: ImageSource.gallery);
+    final File file = File(pickedFile!.path);
+    var duration = Duration(seconds: 1);
+    var image =
+        await ExportVideoFrame.exportImageBySeconds(file, duration, pi / 2);
+>>>>>>> upstream/master
     setState(() {
       widget.images.add(Image.file(image));
       _isClean = true;
     });
+<<<<<<< HEAD
     
+=======
+>>>>>>> upstream/master
     await ExportVideoFrame.saveImage(image, "Video Export Demo",
         waterMark: "images/water_mark.png",
         alignment: Alignment.bottomLeft,
@@ -171,6 +267,7 @@ class _MyAppState extends State<MyApp> {
     }
   }
 
+<<<<<<< HEAD
 
   @override
   Widget build(BuildContext context) {
@@ -184,6 +281,18 @@ class _MyAppState extends State<MyApp> {
             Center(
               child: Text('Running on: $_platformVersion\n'),
             ),
+=======
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Export Image"),
+      ),
+      body: Container(
+        padding: EdgeInsets.zero,
+        child: Column(
+          children: <Widget>[
+>>>>>>> upstream/master
             Expanded(
               flex: 1,
               child: GridView.extent(
@@ -194,8 +303,13 @@ class _MyAppState extends State<MyApp> {
                   crossAxisSpacing: 4,
                   children: widget.images.length > 0
                       ? widget.images
+<<<<<<< HEAD
                       .map((image) => ImageItem(image: image))
                       .toList()
+=======
+                          .map((image) => ImageItem(image: image))
+                          .toList()
+>>>>>>> upstream/master
                       : [Container()]),
             ),
             Expanded(