+++ CHANGELOG.md
@@ -1,3 +1,88 @@
+<<<<<<< HEAD
 ## 0.0.1
 
 * TODO: Describe initial release.
+=======
+# CHANGELOG
+
+## 0.0.1
+
+two interface
+
+* Export Image according to total number image
+* clean cache image
+
+## 0.02
+
+new add two interface
+
+* Export Image according to duration
+* save image to album
+
+### 0.0.2+2
+
+* fix ios export image error sometimes
+
+### 0.0.2+3
+
+* fix some android device get album path error.
+
+### 0.0.2+4
+
+* That save the image to the ablum move to the main thread in android
+
+### 0.0.2+5
+
+* fix bug : after save image to ablum ,but cann't immediately show .
+* That save the image to the ablum recover the sub thread
+
+### 0.0.2+6
+
+* optimize the performance,which export image by number in ios platform
+
+### 0.0.3
+
+* add parameter for export frame.That can configure the quality,rotation of the frame
+
+### 0.0.3+1
+
+* fix rotation image direct problem in ios device.Guarantee rotation is clockwise.
+
+### 0.0.4
+
+add new api
+
+* Export Image frame for gif file
+
+### 0.0.4+1
+
+* Export Image for gif file in Android
+
+### 0.0.5
+
+* Can add watermark when save Image to ablum
+
+### 0.0.5+1
+
+* set watermark position
+
+### 0.0.5+2
+
+* set watermark alignment and scale
+
+### 0.0.5+3
+
+* fix no water mark export fail bug
+
+### 0.0.5+4
+
+* android call back to main thread
+
+### 0.0.6
+
+* fix Null pointer exception in android
+
+### 0.0.7
+
+* support Android embedding v2
+>>>>>>> upstream/master